backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pydantic
openpyxl
python-multipart
